Move "url enrollment" tests to add identity test file
